+++ src/main/java/konkuk/thip/common/security/oauth2/CustomOAuth2UserService.java
@@ -2,7 +2,7 @@ package konkuk.thip.common.security.oauth2;
 
 import konkuk.thip.common.exception.AuthException;
 import konkuk.thip.user.adapter.out.jpa.UserJpaEntity;
-import konkuk.thip.user.adapter.out.persistence.UserJpaRepository;
+import konkuk.thip.user.adapter.out.persistence.repository.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
